[backend] refactor: move pending features to new module
