+++ apps/portal-api/src/modules/common/user-organization.helper.test.ts
@@ -1,8 +1,8 @@
 import { v4 as uuidv4 } from 'uuid';
 import { describe, expect, it } from 'vitest';
 import { PLATFORM_ORGANIZATION_UUID } from '../../portal.const';
+import { UserOrganizationPendingDomain } from '../users/users-pending/user-organization-pending.domain';
 import { createNewUserFromInvitation } from '../users/users.helper';
-import { loadUserOrganizationPending } from './user-organization-pending.domain';
 import { createUserOrganizationRelationAndRemovePending } from './user-organization.helper';
 
 describe('UserOrganizationHelper', () => {
@@ -12,9 +12,10 @@ describe('UserOrganizationHelper', () => {
       const user = await createNewUserFromInvitation({
         email: testMail,
       });
-      const initialPendingOrg = await loadUserOrganizationPending({
-        user_id: user.id,
-      });
+      const initialPendingOrg =
+        await UserOrganizationPendingDomain.loadUserOrganizationPending({
+          user_id: user.id,
+        });
       expect(initialPendingOrg.length).toBe(1);
 
       const user_orgs = await createUserOrganizationRelationAndRemovePending({
@@ -23,9 +24,10 @@ describe('UserOrganizationHelper', () => {
       });
 
       expect(user_orgs.length).toBe(1);
-      const finalPendingOrg = await loadUserOrganizationPending({
-        user_id: user.id,
-      });
+      const finalPendingOrg =
+        await UserOrganizationPendingDomain.loadUserOrganizationPending({
+          user_id: user.id,
+        });
       expect(finalPendingOrg.length).toBe(0);
     });
 
@@ -34,9 +36,10 @@ describe('UserOrganizationHelper', () => {
       const user = await createNewUserFromInvitation({
         email: testMail,
       });
-      const initialPendingOrg = await loadUserOrganizationPending({
-        user_id: user.id,
-      });
+      const initialPendingOrg =
+        await UserOrganizationPendingDomain.loadUserOrganizationPending({
+          user_id: user.id,
+        });
       expect(initialPendingOrg.length).toBe(0);
 
       const user_orgs = await createUserOrganizationRelationAndRemovePending({
