+++ apps/portal-api/src/modules/users/users-pending/user-organization-pending.domain.ts
@@ -1,0 +1,81 @@
+import { db, dbRaw, dbUnsecure, paginate } from '../../../../knexfile';
+import {
+  QueryUsersArgs,
+  UserConnection,
+  User as UserGenerated,
+} from '../../../__generated__/resolvers-types';
+import { requestContext } from '../../../context/request.context';
+import Organization, {
+  OrganizationId,
+} from '../../../model/kanel/public/Organization';
+import User, { UserId } from '../../../model/kanel/public/User';
+import UserOrganizationPending, {
+  UserOrganizationPendingInitializer,
+  UserOrganizationPendingMutator,
+} from '../../../model/kanel/public/UserOrganizationPending';
+
+export const UserOrganizationPendingDomain = {
+  insertNewUserOrganizationPending: (
+    field: UserOrganizationPendingInitializer
+  ): Promise<UserOrganizationPending[]> => {
+    return dbUnsecure<UserOrganizationPending>('User_Organization_Pending')
+      .insert(field)
+      .returning('*');
+  },
+
+  loadOrganizationsWithPendingUsers: async (): Promise<
+    (Organization & { users: User[] })[]
+  > => {
+    return db<Organization>('Organization')
+      .innerJoin(
+        'User_Organization_Pending',
+        'User_Organization_Pending.organization_id',
+        '=',
+        'Organization.id'
+      )
+      .leftJoin('User', 'User.id', '=', 'User_Organization_Pending.user_id')
+      .select('Organization.*', dbRaw('json_agg("User") AS users'))
+      .groupBy('Organization.id');
+  },
+
+  loadPendingUsers: (opts: QueryUsersArgs): Promise<UserConnection> => {
+    const { user } = requestContext.require();
+    const loadPendingUserQuery = db<UserGenerated>('User');
+    loadPendingUserQuery
+      .leftJoin(
+        'User_Organization_Pending as UserOrgPendingFilter',
+        'User.id',
+        'UserOrgPendingFilter.user_id'
+      )
+      .select('User.*')
+      .where(
+        'UserOrgPendingFilter.organization_id',
+        user.selected_organization_id
+      );
+
+    return paginate<UserGenerated, UserConnection>(
+      'User',
+      opts,
+      { unsecured: true },
+      loadPendingUserQuery
+    );
+  },
+
+  loadUserOrganizationPending: (
+    field: UserOrganizationPendingMutator
+  ): Promise<UserOrganizationPending[]> => {
+    return db<UserOrganizationPending>('User_Organization_Pending')
+      .where(field)
+      .secureQuery();
+  },
+
+  removeUserFromOrganizationPending: async (
+    user_id: UserId,
+    organization_id: OrganizationId
+  ) => {
+    return db<UserOrganizationPending>('User_Organization_Pending')
+      .where({ user_id, organization_id })
+      .delete('*')
+      .secureQuery();
+  },
+};
